Refactored the script entry point into a dedicated main module.
'awm.ts' now contains the main class of the library, AdaptiveWebMenus,
which is used by the new init script, 'main.ts'.

This makes it a bit closer to a proper library, even though the design,
naming, etc, of the main class is yet to discuss and improve.

The debug control panel and its styling have also been improved to make
use of this change: adaptation AND policies can now be edited in real
time.