Large refactor of the building logic (now outputs a standalone lib).
This change implies lots of small differences: the build is now much
more generic, as it only produces a ready-to-use library, through the
AdaptiveWebMenus class, exposed in the global scope (see 'awm.ts').

Each website now simply have to include the script importing the AWM
library, and to use specific, short init code to start it up. This is
done using 'js/main.js' script for the test site, and by site-specific
scripts found in the 'examples' directory for some other common websites
(based on their HTML sources).

Configuration files + the makefile have been updated and cleaned to
reflect and make use of those changes!