Large update of the adaptive elements builder methods w/ overload.
Most of the builder have been rewritten in a cleaner, more powerful
ways. It offers new possibilities for user to target menus, groups and
items in simpler ways, with better organised code behind the scene.

In addition, a few other things have been improved, such as a dedicated
value and type for elements built without parent.