Refactored/improved how policies and techniques work and relate.
The main difference is that policies getItemList method now return an
often full list of items (among all), sorted by some criterion depending
on the policy. The technique is now responsible for deciding how many
items to consider in the head of the list, for instance.

In addition, many unused parameters have been removed, so that lists are
more likely to be full, and the code is clearer. Some may be added back
over time. However, techniques may be given more responsability than
before on that topic!