Improved the build process of the library.
The new build structure respects the following logic:

- the bundled, browser-ready (browserified) version of the lib, to be
imported in HTML sources, is now located at the root of the 'build'
directory;

- the CSS file of the lib, to import as well, is also located at the
root of the 'build' directory;

- all the compiled sources are now located in the 'build/lib'
subdirectory, which contains the js entry point + the TS types given to
npm (as stated in the 'package.json' file changes).

This is useful in order to separate the browser bundle (to import in
HTMl sources) from the library sources and their types (to use as a
local linked npm plugin).