Added a fix for reordering of elements split into different parents.
This fix is not very generic, but it avoid the following bug: without an
index correction, a reordered item may not reach the top of its block
(e.g. in a multi-column group of items), as it could be inserted at an
index > 0 while previously reordered elements have been re-inserted in
other blocks!