Refactored the reordering technique, fixing a recently introduced bug.
Reordering now computes reinsertion indices for each group of siblings,
in place for the hacky fix which had been previously added to support
groups were split in several "subgroups". In each group of siblings, it
still uses the indices of adaptive elements of the right type only.

Certain precomputations (e.g. indices) are now done for each group of
siblings (for each call to 'reorderAllElements' method) instead of for
each item, which may also  improve performance for webpages with large
groups/nb of items.