Policies must now include scores along with sorted elements.
Each policy should not only return sorted items, but also augment them
with a "score", which allows to compare how differently were the items
ranked by the policy. This feature is for instance needed for grdual
policies, with different effects applied to the items depending on how
high is their score.

Scores should ideally sum up to 1 over all items with are returned, but
this is probably not always true right now (especially if no items has
stats: in such case, those items are often given a score of 0, thus
summing to 0). This could be part of a future, more formal
specification, but it can hardly make sense for certain policies (e.g.
if items are sorted by the timestamps of the last visit to the page they
point to).

Techniques implementations have been slightly modified in order to
extract ordered items when they only need them.