Added two levels of highlighting (low and high), based on the scores.
A dummy heuristic is currently in use to decide which items should be
highlighted at a high level, but more can be done using the scores of
the items (which are now returned by any item list policy).

The difference between high and low levels of highlighting is the
opacity (twice as high for the high level than for the low level); see
CSS file for details.