Refactoring of the technique & policy system.
Policies now all inherit from a 'Policy' abstract class (instead of
implementing an interface), which provides both soretd item and group
methods. The item sorting and scoring method implementation is left to
concrete child classes, but a default group sorting and scoring method
is provided (based on item score sums, using the aformentioned related
method).

Additional methods to only get the sorted items and groups are also
provided.

Techniques keep implementing an interface, renamed to 'Technique' and
moved to its own file.

Not only this allows all policies to provide item and group
sorting/scoring with less duplicated code, but it could allow to
implement features for all policies at once, in a generic way (e.g. for
emitting events).

The same approach could be taken for techniques (move from an interface
to an abstract class) at a later stage, if need be.