Added a first draft of the progressive HL + RI policy.
It needs further testing and improvements (e.g. score thresholds to 
switch between states), and does not produce any good results right now, 
but the basics are working.

Part of the library have been slightly adapted for this addition, as 
this technique requires to save some history about previous adaptations 
in a persistent way, and requires an access to the database dedicated 
object for this purpose.