Renamed the database persistent storage + more robust database init.
The database is now always initialized with default values (i.e. empty 
database), and loads whatever it can from the local storage. This makes 
it easier to change the structure of what is saved:loaded by the 
database in the local storage (as it has been the case when renaming the 
persistent storage, hence this combined addition).