Fixed index sorting, done in lexicographical order by default.
Array.sort casts non-undefined elements to string before comparing them 
in lexicographical order, unless a comparator is given.

Reordering indices were thus wrongly sorted, but that would only become 
visible when 3+ items of the same group were rordered (because indices 1 
and 2 were sorted above 10 and 11, and a group had to be 9+ items long 
for 3+ items to be reordered).