Updated some dependencies.
Note: @types/jquery should have been updated to version 3.3.4, but 
versions 3.3.3 and above seem to introduce a change in what types are 
accepted by the static $ constructor, breaking the use of AWM's Selector 
type. This is why the version number has been set to 3.3.2 specifically 
(this should be updated in the future, if the issue is solved, or 
directly in AWM).