Added a data manager, which handles the database, logger and analayser.
This new class serves as a unique entry point for the "data" module of 
the library. It takes the responsability of everything related with data 
in this library: logging it, saving or loading it, as well as providing 
analysis tools.

This refactoring not only better split the responsability and instance 
ownership, but also gives more flexibility to the policies, which can 
seamlessly readfrom or write into the database, avoiding dirty 
workarounds which have been previously used!