Added a menu manager, which handles all adaptive menus.
In a similar fashion to the last refactoring, this one delegates all the 
responsabilities for handling the menus (and other menu elements such as 
groups and items) to a new menu manager.

All classes which used to be given an array of menus are now given a 
reference to the menu manager, which offers several methods to access 
the menus and their content.