Added code style fixes and improvements, added for...in loop checks.
When chaining method calls, each call should be put on a new line. 
However, I did not follow this rule when chaining a function call and a 
single method call, e.g. when building a jQuery object, and calling a 
metod on it:

$(foo).bar(baz);

Looks fine, and does not need to be turned into:

$(foo)
  .bar(baz);
  
in my opinion (as long as it looks clean and clear enough). However,

someObject
  method1()
  method2()
  method3();

is now prefered, and considered standard in AWM soures.

Moreover, the remaining for...in loops have been improved with 
hasOwnProperty standard checks, as it is a good practice, whose lack was 
actually flagged by tslint.