Moved available techniques and policies to the adaptation manager.
Related list of names and types have been moved as well.

Since the adaptation manager was almost the only class using those 
values and types, and has the responsability of handling the adaptations 
and exposing them to the rest of the library, it made sense to directly 
define and instanciate them (as sort-of singletons) there.