Removed the useless type parameter of Technique interface.
The initial goal of this type parameter was to be able to check if 
techniques were only given compatible policies at compile time; but it 
had no effect in practice: (1) switching between techniques and policies 
is mostly a runtime behaviour, and (2) the latter are often decribed by 
their names, i.e. strings, which do not allow to easily check such 
constraints.