Shorten some of the single-line doc comments.
Typedoc appears to handle single-line documentation comments, as long as they start by `/**`.

Therefore, this commit removes the two unecessary line breaks for (1) class properties and (2) any object defined outside of a class.

Comments of other class members have not be shortened to ensure methods are sufficiently spaced, and for easy extension of their documentation.